Expose native Windows GNU regressions
Add behavior-level coverage for the native Windows GNU Zig target, unread terminal probes, and Windows drive file URLs while preserving the pre-fix behavior.\n\nConstraint: cmux-tui hosted validation is required; local Rust, Cargo, and Zig execution is prohibited by cmux-tui/AGENTS.md.\nRejected: source-only assertions | these tests exercise the target-selection helper and runtime URL/probe seams.\nConfidence: high\nScope-risk: narrow\nDirective: Keep the implementation in a follow-up commit so CI can prove these regressions are real.\nTested: git diff --check\nNot-tested: Rust tests and hosted CI; this commit intentionally precedes the fix.,
